+++ src/main/java/gtclassic/api/helpers/GTValues.java
@@ -38,14 +38,15 @@ public class GTValues {
 			CIRCUIT_ULTIMATE = "circuitUltimate", CHEST_WOOD = "chestWood", DUCT_TAPE = "craftingToolDuctTape",
 			DUST_GLOWSTONE = "dustGlowstone", INGOT_PLATINUM = "ingotPlatinum", INGOT_BRONZE = "ingotBronze",
 			INGOT_SILICON = "itemSilicon", INGOT_ELECTRUM = "ingotElectrum", INGOT_SILVER = "ingotSilver",
-			INGOT_INVAR = "ingotInvar", INGOT_NICKEL = "ingotNickel", INGOT_GOLD = "ingotGold";
+			INGOT_INVAR = "ingotInvar", INGOT_NICKEL = "ingotNickel", INGOT_GOLD = "ingotGold",
+			ENDER_CHEST = "chestEnder";
 	/** Generic fluidstack holders **/
 	public static final FluidStack FS_WATER = new FluidStack(FluidRegistry.WATER, 1000),
 			FS_LAVA = new FluidStack(FluidRegistry.LAVA, 1000);
 	/** Stores combine recipe input vars **/
 	public static final IRecipeInput INPUT_INGOT_MACHINE = GTRecipeCraftingHandler.combineRecipeObjects(INGOT_REFINEDIRON, INGOT_ALUMINIUM),
 			INPUT_INGOT_ELECTRIC = GTRecipeCraftingHandler.combineRecipeObjects(INGOT_REFINEDIRON, INGOT_SILICON, INGOT_ALUMINIUM, INGOT_SILVER, INGOT_ELECTRUM, INGOT_PLATINUM, INGOT_GOLD),
-			INPUT_INGOT_ANY = GTRecipeCraftingHandler.combineRecipeObjects("ingotIron", "ingotRefinedIron", "ingotSteel", INGOT_SILVER, INGOT_BRONZE, INGOT_ALUMINIUM, INGOT_ELECTRUM, INGOT_PLATINUM, INGOT_NICKEL, INGOT_INVAR),
+			INPUT_INGOT_ANY = GTRecipeCraftingHandler.combineRecipeObjects("ingotIron", "ingotRefinedIron", "ingotSteel", INGOT_SILVER, INGOT_BRONZE, INGOT_ALUMINIUM, INGOT_ELECTRUM, INGOT_PLATINUM, INGOT_NICKEL, INGOT_INVAR, INGOT_TITANIUM, INGOT_TUNGSTEN, INGOT_CHROME),
 			INPUT_INGOT_ANYIRON = GTRecipeCraftingHandler.combineRecipeObjects("ingotRefinedIron", "ingotIron", "ingotSteel"),
 			INPUT_INGOT_MIXED = GTRecipeCraftingHandler.combineRecipeObjects(INGOT_SILVER, INGOT_ALUMINIUM, INGOT_ELECTRUM, INGOT_INVAR),
 			INPUT_CRYSTAL_LOW = GTRecipeCraftingHandler.combineRecipeObjects(GEM_DIAMOND, "gemRuby"),